CSPACE-3070: Fixed several existing boundary condition errors that were revealed by the previous commit -see http://fisheye.collectionspace.org/changelog/collectionspace/?cs=3415

+++ services/authorization-mgt/service/src/main/java/org/collectionspace/services/authorization/storage/AuthorizationDelegate.java
@@ -118,42 +118,48 @@ public class AuthorizationDelegate {
         SubjectType subject = PermissionRoleUtil.getRelationSubject(ctx, pr);
         AuthZ authz = AuthZ.get();
         if (subject.equals(SubjectType.ROLE)) {
-            PermissionValue pv = pr.getPermissions().get(0);
-            Permission p = getPermission(pv.getPermissionId());
-            if (p == null) {
-                String msg = "deletePermissions: No permission found for id=" + pv.getPermissionId();
-                logger.error(msg);
-                throw new DocumentNotFoundException(msg);
-            }
-            CSpaceResource[] resources = getResources(p);
-            String[] roles = getRoles(pr.getRoles());
-            for (CSpaceResource res : resources) {
-                authz.deletePermissions(res, roles);
-            }
+        	List<PermissionValue> permissionValues = pr.getPermissions();
+        	if (permissionValues != null & permissionValues.size() > 0) {
+	            PermissionValue pv = permissionValues.get(0);
+	            Permission p = getPermission(pv.getPermissionId());
+	            if (p == null) {
+	                String msg = "deletePermissions: No permission found for id=" + pv.getPermissionId();
+	                logger.error(msg);
+	                throw new DocumentNotFoundException(msg);
+	            }
+	            CSpaceResource[] resources = getResources(p);
+	            String[] roles = getRoles(pr.getRoles());
+	            for (CSpaceResource res : resources) {
+	                authz.deletePermissions(res, roles);
+	            }
+        	}
         } else if (SubjectType.PERMISSION.equals(subject)) {
-            RoleValue rv = pr.getRoles().get(0);
-            Role r = getRole(rv.getRoleId());
-            if (r == null) {
-                String msg = "deletePermissions: No role found for id=" + rv.getRoleId();
-                logger.error(msg);
-                throw new DocumentNotFoundException(msg);
-            }
-            //using r not rv ensures we're getting the "ROLE" prefix/qualified name
-            // This needs to use the qualified name, not the display name
-            String[] roles = {r.getRoleName()}; 
-            for (PermissionValue pv : pr.getPermissions()) {
-                Permission p = getPermission(pv.getPermissionId());
-                if (p == null) {
-                    String msg = "deletePermissions: No permission found for id=" + pv.getPermissionId();
-                    logger.error(msg);
-                    //TODO: would be nice contiue to still send 400 back
-                    continue;
-                }
-                CSpaceResource[] resources = getResources(p);
-                for (CSpaceResource res : resources) {
-                    authz.deletePermissions(res, roles);
-                }
-            }
+        	List<RoleValue> roleValues = pr.getRoles();
+        	if (roleValues != null && roleValues.size() > 0) {
+	            RoleValue rv = roleValues.get(0);
+	            Role r = getRole(rv.getRoleId());
+	            if (r == null) {
+	                String msg = "deletePermissions: No role found for id=" + rv.getRoleId();
+	                logger.error(msg);
+	                throw new DocumentNotFoundException(msg);
+	            }
+	            //using r not rv ensures we're getting the "ROLE" prefix/qualified name
+	            // This needs to use the qualified name, not the display name
+	            String[] roles = {r.getRoleName()}; 
+	            for (PermissionValue pv : pr.getPermissions()) {
+	                Permission p = getPermission(pv.getPermissionId());
+	                if (p == null) {
+	                    String msg = "deletePermissions: No permission found for id=" + pv.getPermissionId();
+	                    logger.error(msg);
+	                    //TODO: would be nice contiue to still send 400 back
+	                    continue;
+	                }
+	                CSpaceResource[] resources = getResources(p);
+	                for (CSpaceResource res : resources) {
+	                    authz.deletePermissions(res, roles);
+	                }
+	            }
+        	}
         }
     }
